ai-backend/requirements.txt
Flask==2.0.3
torch==1.13.1
torchvision==0.14.1
Pillow==8.4.0
opencv-python-headless==4.5.3.56
Werkzeug==2.0.3
NumPy==1.23.5  # Add this line to downgrade NumPy
